Feature-scheduler (#20)
* 🚀 feat(setup.py): add new dependencies requests and sqlalchemy
🚀 feat(scheduler.py): add new module for scheduling trading engine updates and stats retrieval
🔥 chore(telegram_interface): remove unused file
🐛 fix(web_manager.py): add logging for closing position due to price being outside of range
The setup.py file now includes two new dependencies, requests and sqlalchemy. The scheduler.py file is a new module that schedules the trading engine updates and stats retrieval. The telegram_interface/__main__.py file was removed as it was not being used. The web_manager.py file now includes logging for closing a position due to the price being outside of the range.

* 🐛 fix(docker-compose.yml): add uniswap-hft-scheduler service
✨ feat(docker-compose.yml): add uniswap-hft-scheduler service to docker-compose file
The uniswap-hft-scheduler service was added to the docker-compose file. This service runs the uniswap_hft.scheduler.scheduler module, which is responsible for scheduling the execution of the trading bot. The service is dependent on the uniswap-hft-api service, which provides the necessary data for the scheduler to make trading decisions.

* 🖤 Format code with black

---------

Co-authored-by: adradr <adradr@users.noreply.github.com>